+++ sources/Google.Solutions.IapDesktop.Extensions.Session.Test/Protocol/Adapter/TestKeyStoreAdapter.cs
@@ -21,7 +21,7 @@
 
 using Google.Solutions.Apis.Auth;
 using Google.Solutions.IapDesktop.Extensions.Session.Protocol.Adapter;
-using Google.Solutions.Ssh.Auth;
+using Google.Solutions.Ssh.Cryptography;
 using Moq;
 using NUnit.Framework;
 using System;
