+++ sources/Google.Solutions.IapDesktop.Extensions.Session.Test/Protocol/Ssh/TestSessionContextFactoryForSsh.cs
@@ -33,7 +33,7 @@ using Google.Solutions.IapDesktop.Extensions.Session.Protocol.Rdp;
 using Google.Solutions.IapDesktop.Extensions.Session.Protocol.Ssh;
 using Google.Solutions.IapDesktop.Extensions.Session.Settings;
 using Google.Solutions.IapDesktop.Extensions.Session.ToolWindows.Credentials;
-using Google.Solutions.Ssh.Auth;
+using Google.Solutions.Ssh.Cryptography;
 using Microsoft.Win32;
 using Moq;
 using NUnit.Framework;
