+++ sources/Google.Solutions.IapDesktop.Extensions.Session.Test/ToolWindows/Options/TestSshOptionsViewModel.cs
@@ -23,7 +23,7 @@ using Google.Solutions.Common.Util;
 using Google.Solutions.IapDesktop.Application.Profile;
 using Google.Solutions.IapDesktop.Extensions.Session.Settings;
 using Google.Solutions.IapDesktop.Extensions.Session.ToolWindows.Options;
-using Google.Solutions.Ssh.Auth;
+using Google.Solutions.Ssh.Cryptography;
 using Microsoft.Win32;
 using NUnit.Framework;
 using System;
