+++ sources/Google.Solutions.IapDesktop.Extensions.Session.Test/ToolWindows/Ssh/TestSshTerminalPaneViewModel.cs
@@ -33,7 +33,7 @@ using Google.Solutions.IapDesktop.Extensions.Session.ToolWindows.Ssh;
 using Google.Solutions.Mvvm.Controls;
 using Google.Solutions.Platform.Security;
 using Google.Solutions.Ssh;
-using Google.Solutions.Ssh.Auth;
+using Google.Solutions.Ssh.Cryptography;
 using Google.Solutions.Testing.Apis.Integration;
 using Google.Solutions.Testing.Application.Mocks;
 using Moq;
