+++ sources/Google.Solutions.IapDesktop.Extensions.Session/Protocol/Ssh/KeyAuthorizer.cs
@@ -27,7 +27,7 @@ using Google.Solutions.Common.Diagnostics;
 using Google.Solutions.Common.Util;
 using Google.Solutions.IapDesktop.Application;
 using Google.Solutions.IapDesktop.Core.ObjectModel;
-using Google.Solutions.Ssh.Auth;
+using Google.Solutions.Ssh.Cryptography;
 using System;
 using System.Threading;
 using System.Threading.Tasks;
