+++ sources/Google.Solutions.IapDesktop.Extensions.Session/Protocol/Ssh/OsLoginProfile.cs
@@ -28,7 +28,7 @@ using Google.Solutions.Common.Util;
 using Google.Solutions.IapDesktop.Application;
 using Google.Solutions.IapDesktop.Application.Host.Adapters;
 using Google.Solutions.IapDesktop.Core.ObjectModel;
-using Google.Solutions.Ssh.Auth;
+using Google.Solutions.Ssh.Cryptography;
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
