+++ sources/Google.Solutions.IapDesktop.Extensions.Session/Settings/SshSettings.cs
@@ -25,7 +25,7 @@ using Google.Solutions.IapDesktop.Application.Profile.Settings;
 using Google.Solutions.IapDesktop.Application.Profile.Settings.Registry;
 using Google.Solutions.IapDesktop.Core.ObjectModel;
 using Google.Solutions.IapDesktop.Extensions.Session.Protocol.Ssh;
-using Google.Solutions.Ssh.Auth;
+using Google.Solutions.Ssh.Cryptography;
 using Microsoft.Win32;
 using System;
 using System.Collections.Generic;
