+++ sources/Google.Solutions.IapDesktop.Extensions.Session/ToolWindows/Options/SshOptionsSheet.cs
@@ -23,7 +23,7 @@ using Google.Solutions.Common.Diagnostics;
 using Google.Solutions.Common.Util;
 using Google.Solutions.IapDesktop.Core.ObjectModel;
 using Google.Solutions.Mvvm.Binding;
-using Google.Solutions.Ssh.Auth;
+using Google.Solutions.Ssh.Cryptography;
 using System;
 using System.ComponentModel.DataAnnotations;
 using System.Linq;
