+++ sources/Google.Solutions.Ssh/Native/ISshAuthenticator.cs
@@ -20,7 +20,7 @@
 //
 
 using Google.Solutions.Common.Util;
-using Google.Solutions.Ssh.Auth;
+using Google.Solutions.Ssh.Cryptography;
 using System;
 
 namespace Google.Solutions.Ssh.Native
